% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_evenness.R
\name{shannon_evenness}
\alias{shannon_evenness}
\title{Shannon Index example}
\usage{
shannon_evenness(nSim = 1000, nspV = 10:20)
}
\arguments{
\item{nSim}{numeric. Number of simulations}

\item{nspV}{numeric vector. Number of entities to simulate}
}
\description{
Demonstrate some of the issues in interpreting changes in the index
Evenness in p's
}
