% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_skewed.R
\name{shannon_skewed}
\alias{shannon_skewed}
\title{Shannon Index example}
\usage{
shannon_skewed(nSim = 1000, nspV = 10:20)
}
\arguments{
\item{nSim}{numeric. Number of simulations}

\item{nspV}{numeric vector. Number of entities to simulate}
}
\description{
Demonstrate some of the issues in interpreting changes in the index
}
\details{
Sampling ps sequentially from uniform subject to sum(p) = 1
Skews to top heavy populations
}
